refactor: rename neww component
