+++ src/frontend/src/lib/components/core/Footer.svelte
@@ -2,8 +2,8 @@
 	import { IconGitHub } from '@dfinity/gix-components';
 	import IconDfinity from '$lib/components/icons/IconDfinity.svelte';
 	import IconTwitter from '$lib/components/icons/IconTwitter.svelte';
-	import ButtonAsLink from '$lib/components/ui/ButtonAsLink.svelte';
 	import ExternalLink from '$lib/components/ui/ExternalLink.svelte';
+	import ExternalLinkIcon from '$lib/components/ui/ExternalLinkIcon.svelte';
 	import { OISY_REPO_URL } from '$lib/constants/oisy.constants';
 	import { authNotSignedIn } from '$lib/derived/auth.derived';
 	import { i18n } from '$lib/stores/i18n.store';
@@ -20,16 +20,16 @@
 	class:md:h-md:w-[50%]={$authNotSignedIn}
 >
 	<div class="flex flex-row items-center gap-4">
-		<ButtonAsLink href={OISY_REPO_URL} ariaLabel={$i18n.navigation.text.source_code_on_github}>
+		<ExternalLinkIcon href={OISY_REPO_URL} ariaLabel={$i18n.navigation.text.source_code_on_github}>
 			<IconGitHub />
-		</ButtonAsLink>
+		</ExternalLinkIcon>
 
-		<ButtonAsLink
+		<ExternalLinkIcon
 			href={OISY_TWITTER_URL}
 			ariaLabel={replaceOisyPlaceholders($i18n.navigation.alt.open_twitter)}
 		>
 			<IconTwitter />
-		</ButtonAsLink>
+		</ExternalLinkIcon>
 	</div>
 
 	<div class="flex flex-row items-center gap-2 text-sm">
