+++ src/frontend/src/lib/components/ui/ExternalLinkIcon.svelte
